Add clang-tidy with bug-finding checks and fix all warnings
Add .clang-tidy config with bugprone-*, clang-analyzer-*, and
performance-* checks. Suppress vendored code (third_party/, cjson/,
stb_truetype.h) and noisy false-positive checks for Vulkan/Wayland
API patterns. CMake integration via -DENABLE_CLANG_TIDY=ON.

Fixes found by clang-tidy:
- single_instance: close() clobbering errno before check
- main: remove dead variables video_ready and current_playback_rate
- mpv_player: std::move callbacks instead of copying std::function
- cef_client/menu_overlay: fix implicit int widening in buffer sizes
- cef_app: use !empty() instead of size() >= 1
- cef_thread/mpv_player/single_instance: pass large params by const ref
- browser_layer/sdl_to_vk: add default cases, merge identical branches
- opengl_renderer/wayland_subsurface/x11_video_layer: qualify virtual
  calls in destructors to suppress dispatch warning

+++ src/cef/cef_client.cpp
@@ -482,7 +482,7 @@ void Client::OnPaint(CefRefPtr<CefBrowser> browser, PaintElementType type,
     if (!on_paint_) return;
 
     if (type == PET_POPUP) {
-        size_t size = width * height * 4;
+        size_t size = static_cast<size_t>(width) * height * 4;
         popup_buffer_.resize(size);
         memcpy(popup_buffer_.data(), buffer, size);
         popup_pixel_width_ = width;
@@ -502,7 +502,7 @@ void Client::OnPaint(CefRefPtr<CefBrowser> browser, PaintElementType type,
     }
 
     // Slow path: blend popup onto view (only when dropdown is visible)
-    size_t size = width * height * 4;
+    size_t size = static_cast<size_t>(width) * height * 4;
     composite_buffer_.resize(size);
     memcpy(composite_buffer_.data(), buffer, size);
